enhance: refine variable-length-type memory usage(#38736) (#43093)
related: #38736
pr: https://github.com/milvus-io/milvus/pull/39578

+++ internal/core/src/common/Common.h
@@ -67,7 +67,7 @@ SetDefaultConfigParamTypeCheck(bool val);
 struct BufferView {
     struct Element {
         const char* data_;
-        uint64_t* offsets_;
+        uint32_t* offsets_;
         int start_;
         int end_;
     };
